fix(ci): move audit-dashboard route tests out of test_dashboards.py
Slim fastlane planned the whole tests/test_dashboards.py module because
the new #7080 audit-dashboard route tests were added there, pulling in
the unmapped third-party import generate_playground_data. Move the tests
to dedicated tests/test_audit_dashboard_routes.py and restore
test_dashboards.py to origin/main.

X-Agent: kimi/infra-7080-fastlane
